修改所有 customerId为tabCustomerId

+++ src/main/java/com/atat/customer/service/TabCustomerService.java
@@ -46,13 +46,13 @@ public interface TabCustomerService {
      * @param tabCustomerId
      * @return
      */
-    public Map<String, Object> getTabCustomerById(Integer tabCustomerId);
+    public Map<String, Object> getTabCustomerById(Long tabCustomerId);
 
     /**
      * 依据Id删除 用户表 记录
      * @param tabCustomerId
      */
-    public void delTabCustomerById(Integer tabCustomerId);
+    public void delTabCustomerById(Long tabCustomerId);
 
     /**
      * 依据微信Id获取用户
